requirements.txt
psycopg2-binary ==2.8.5
pymysql~=0.9.3
pyyaml ==5.3
requests==2.24.0


pygame~=2.1.2